Added a separate test target to run Nat test cases[#28773663]

+++ src/maidsafe/routing/tests/routing_network.cc
@@ -1119,7 +1119,7 @@ testing::AssertionResult GenericNetwork::SendDirect(const size_t& repeats) {
   }
 
   uint16_t count(0);
-  uint16_t max_count(30 * (nodes_.size()) * (nodes_.size() - 1));
+  uint16_t max_count(static_cast<uint16_t>(30 * (nodes_.size()) * (nodes_.size() - 1)));
   while (reply_count < repeats * total_num_nodes * total_num_nodes) {
     ++count;
     if (count == max_count) {
@@ -1207,7 +1207,7 @@ testing::AssertionResult GenericNetwork::SendGroup(const NodeId& target_id,
   }
 
   uint16_t count(0);
-  uint16_t max_count(300 * repeats);
+  uint16_t max_count(static_cast<uint16_t>(300 * repeats));
   while (reply_count < Parameters::node_group_size * repeats) {
     ++count;
     if (count == max_count) {
